requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning & RL
torch>=2.0.0
stable-baselines3>=2.0.0
gym>=0.26.0
tensorboard>=2.13.0

# Financial Data
yfinance>=0.2.28
pandas-datareader>=0.10.0
finnhub-python>=2.4.0
alpha-vantage>=2.3.1

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14b0
technical-indicators>=0.1.0

# Streamlit Dashboard
streamlit>=1.28.0
plotly>=5.17.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6

# Trading APIs (Optional - for paper/live trading)
alpaca-py>=0.11.0
alpaca-trade-api>=3.0.0

# Data Processing
scikit-learn>=1.3.0
joblib>=1.3.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Logging & Monitoring
loguru>=0.7.0
python-telegram-bot>=20.0  # Optional: pro notifikace

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Scheduling (backup if not using GitHub Actions)
schedule>=1.2.0
APScheduler>=3.10.0
